crash bug

+++ src/test.py
@@ -2,6 +2,7 @@
 import urllib2
 import urllib
 import re
+import os
 
 start_url = 'http://girl-atlas.com/'
 
@@ -13,7 +14,12 @@ def read_file():
 
 def  get_html_content(url):
 	print "fecthing:" + url
-	return urllib2.urlopen(url).read()
+	content = ''
+	try:
+		content = urllib2.urlopen(url).read()
+	except:
+		print "get html error!"
+	return content
 
 def get_jpg_list(content):
 	reg = r'(?:\'|\")http[^ ]+?(?:jpg!mid|jpg|png)(?:\'|\")'
@@ -29,11 +35,20 @@ def get_url_list(content):
 
 def download_file(uri):	
 	l = uri.split('/')
-	name = '/home/yue/workspace/python/img/' + l[len(l) - 1]	
-	f = urllib2.urlopen(uri) 
-	data = f.read() 
-	with open(name, "wb") as code:
-		code.write(data)
+	name = '/home/yue/workspace/python/img/' + l[len(l) - 1]
+
+	if os.path.exists(name):
+		return
+		
+	try:
+		f = urllib2.urlopen(uri)
+		data = f.read() 
+		with open(name, "wb") as code:
+			code.write(data)
+	except Exception, e:
+		print "get error while download!"
+	
+	
 
 def is_except_str(content):
 	if len(content) < 14:
@@ -48,11 +63,21 @@ def is_except_str(content):
 
 
 to_fetch_urls = [start_url]
+has_fetch_urls = []
 
 while to_fetch_urls: 
-	html = get_html_content(to_fetch_urls[0])
+	html = get_html_content(to_fetch_urls[0])	
+	
+	if len(html) < 5:
+		temp_url = to_fetch_urls[0]
+		del to_fetch_urls[0]
+		to_fetch_urls.append(temp_url)
+		continue
+
 	print("fecth url success!")
+	has_fetch_urls.append(to_fetch_urls[0])
 	del to_fetch_urls[0]
+	
 
 	jpg_list = get_jpg_list(html)
 	for jpg in jpg_list:
@@ -66,7 +91,7 @@ while to_fetch_urls:
 	url_list = get_url_list(html)	
 	for url in url_list:		
 		if not is_except_str(url):
-			if url not in to_fetch_urls:
+			if url not in to_fetch_urls and url not in has_fetch_urls:
 				url = url.strip('\"')
 				url = url.strip('\'')
 				to_fetch_urls.append(url)
